requirements.txt
appnope==0.1.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
certifi==2016.9.26
configparser==3.5.0
decorator==4.0.10
entrypoints==0.2.2
enum34==1.1.6
functools32==3.2.3.post2
ipykernel==4.5.2
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.1
MarkupSafe==0.23
mistune==0.7.3
nbconvert==4.2.0
nbformat==4.2.0
notebook==4.3.0
numpy==1.11.2
pandas==0.19.1
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.6.0
python-slugify==1.2.1
pytz==2016.10
pyzmq==16.0.2
qtconsole==4.2.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
terminado==0.6
tornado==4.4.2
traitlets==4.3.1
Unidecode==0.4.19
wcwidth==0.1.7
widgetsnbextension==1.2.6
xlrd==1.0.0
